remove partial check (it doesn't matter)
if the message being reacted to (for example, "react with a ✅ to
continue") does not exist in Discord.js's cache (for
example, if this app's process has been restarted), only a
partial `messageReaction` will be received.

a partial `messageReaction` is indicated by the
`messageReaction.partial`
property. it will contain some or _part_ of the data.

we need `messageReaction.emoji.name` and
`messageReaction.message.channel` to execute the event handler
successfully. in my testing, these properties have always been available
even if `messageReaction.partial` equals `true`

it may be that `partial` messages contain _different_ `parts` of the
data depending on curcumstnaces, which could lead to an exception.

if we discover that to be true, it should be possible to fetch the
missing message data (for example, `if (messageReaction.partial) {
messageReaction = await messageReaction.fetch() }` (i haven't looked up
the precise function and it's usage)